back/requirements.txt
asgiref==3.8.1
astroid==3.1.0
black==24.3.0
coverage
click==8.1.7
dill==0.3.8
Django>=5.0.3,<5.1
django-dotenv
django-storages==1.11.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary
pylint==3.1.0
pytest==8.1.1
pytest-django==4.8.0
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.10.0
boto3
requests
gunicorn